[core] Extract an SST File Format from LookupStore.
