+++ paimon-common/src/main/java/org/apache/paimon/sst/BlockEntry.java
@@ -16,35 +16,40 @@
  * limitations under the License.
  */
 
-package org.apache.paimon.lookup.sort;
+package org.apache.paimon.sst;
 
-import java.util.Objects;
+import org.apache.paimon.memory.MemorySlice;
 
-/** Handle for bloom filter. */
-public class BloomFilterHandle {
+import java.util.Map.Entry;
 
-    public static final int MAX_ENCODED_LENGTH = 9 + 5 + 9;
+import static java.util.Objects.requireNonNull;
 
-    private final long offset;
-    private final int size;
-    private final long expectedEntries;
+/** Entry represents a key value. */
+public class BlockEntry implements Entry<MemorySlice, MemorySlice> {
 
-    BloomFilterHandle(long offset, int size, long expectedEntries) {
-        this.offset = offset;
-        this.size = size;
-        this.expectedEntries = expectedEntries;
+    private final MemorySlice key;
+    private final MemorySlice value;
+
+    public BlockEntry(MemorySlice key, MemorySlice value) {
+        requireNonNull(key, "key is null");
+        requireNonNull(value, "value is null");
+        this.key = key;
+        this.value = value;
     }
 
-    public long offset() {
-        return offset;
+    @Override
+    public MemorySlice getKey() {
+        return key;
     }
 
-    public int size() {
-        return size;
+    @Override
+    public MemorySlice getValue() {
+        return value;
     }
 
-    public long expectedEntries() {
-        return expectedEntries;
+    @Override
+    public final MemorySlice setValue(MemorySlice value) {
+        throw new UnsupportedOperationException();
     }
 
     @Override
@@ -55,14 +60,19 @@ public class BloomFilterHandle {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        BloomFilterHandle that = (BloomFilterHandle) o;
-        return offset == that.offset
-                && size == that.size
-                && expectedEntries == that.expectedEntries;
+
+        BlockEntry entry = (BlockEntry) o;
+
+        if (!key.equals(entry.key)) {
+            return false;
+        }
+        return value.equals(entry.value);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(offset, size, expectedEntries);
+        int result = key.hashCode();
+        result = 31 * result + value.hashCode();
+        return result;
     }
 }
