+++ paimon-common/src/main/java/org/apache/paimon/sst/BlockTrailer.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.paimon.lookup.sort;
+package org.apache.paimon.sst;
 
 import org.apache.paimon.compression.BlockCompressionType;
 import org.apache.paimon.memory.MemorySlice;
