+++ paimon-common/src/main/java/org/apache/paimon/utils/FileBasedBloomFilter.java
@@ -19,7 +19,8 @@
 package org.apache.paimon.utils;
 
 import org.apache.paimon.annotation.VisibleForTesting;
-import org.apache.paimon.io.PageFileInput;
+import org.apache.paimon.fs.Path;
+import org.apache.paimon.fs.SeekableInputStream;
 import org.apache.paimon.io.cache.CacheCallback;
 import org.apache.paimon.io.cache.CacheKey;
 import org.apache.paimon.io.cache.CacheManager;
@@ -34,16 +35,18 @@ import static org.apache.paimon.utils.Preconditions.checkArgument;
 /** Util to apply a built bloom filter . */
 public class FileBasedBloomFilter implements Closeable {
 
-    private final PageFileInput input;
+    private final SeekableInputStream input;
     private final CacheManager cacheManager;
     private final BloomFilter filter;
     private final long readOffset;
-    private final int readLength;
     private final CacheKey cacheKey;
+    // each bloom filter is only used by a single file reader, so we can safely reuse here
+    private final byte[] reusedPageBuffer;
     private int accessCount;
 
     public FileBasedBloomFilter(
-            PageFileInput input,
+            SeekableInputStream input,
+            Path filePath,
             CacheManager cacheManager,
             long expectedEntries,
             long readOffset,
@@ -53,9 +56,9 @@ public class FileBasedBloomFilter implements Closeable {
         checkArgument(expectedEntries >= 0);
         this.filter = new BloomFilter(expectedEntries, readLength);
         this.readOffset = readOffset;
-        this.readLength = readLength;
         this.accessCount = 0;
-        this.cacheKey = CacheKey.forPosition(input.file(), readOffset, readLength, true);
+        this.cacheKey = CacheKey.forPosition(filePath, readOffset, readLength, true);
+        this.reusedPageBuffer = new byte[readLength];
     }
 
     public boolean testHash(int hash) {
@@ -65,13 +68,17 @@ public class FileBasedBloomFilter implements Closeable {
         if (accessCount == REFRESH_COUNT || filter.getMemorySegment() == null) {
             MemorySegment segment =
                     cacheManager.getPage(
-                            cacheKey,
-                            key -> input.readPosition(readOffset, readLength),
-                            new BloomFilterCallBack(filter));
+                            cacheKey, key -> readPage(), new BloomFilterCallBack(filter));
             filter.setMemorySegment(segment, 0);
             accessCount = 0;
         }
         return filter.testHash(hash);
+    }
+
+    private byte[] readPage() throws IOException {
+        input.seek(readOffset);
+        IOUtils.readFully(input, reusedPageBuffer);
+        return reusedPageBuffer;
     }
 
     @VisibleForTesting
